Renamed FAISS related files and classes
